Move ListBoxEx and RichTextBoxEx to AppTemplate.
